+++ app/src/main/java/com/techcourse/TomcatStarter.java
@@ -1,13 +1,13 @@
 package com.techcourse;
 
+import com.techcourse.servlet.UncheckedServletException;
+import java.io.File;
 import org.apache.catalina.Context;
 import org.apache.catalina.LifecycleException;
 import org.apache.catalina.connector.Connector;
 import org.apache.catalina.core.StandardContext;
 import org.apache.catalina.startup.Tomcat;
 import org.apache.tomcat.util.scan.StandardJarScanner;
-
-import java.io.File;
 
 public class TomcatStarter {
 
@@ -32,7 +32,7 @@ public class TomcatStarter {
     public void start() {
         try {
             tomcat.start();
-        } catch (LifecycleException e) {
+        } catch (final LifecycleException e) {
             throw new UncheckedServletException(e);
         }
     }
@@ -41,7 +41,7 @@ public class TomcatStarter {
         try {
             tomcat.stop();
             tomcat.destroy();
-        } catch (LifecycleException e) {
+        } catch (final LifecycleException e) {
             throw new UncheckedServletException(e);
         }
     }
