+++ app/src/main/java/com/techcourse/servlet/DispatcherServletInitializer.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.servlet;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
@@ -6,8 +6,8 @@ import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
 
 /**
- * Base class for {@link WebApplicationInitializer}
- * implementations that register a {@link DispatcherServlet} in the servlet context.
+ * Base class for {@link WebApplicationInitializer} implementations that register a {@link DispatcherServlet} in the
+ * servlet context.
  */
 public class DispatcherServletInitializer implements WebApplicationInitializer {
