+++ app/src/main/java/com/techcourse/servlet/ManualHandlerMapping.java
@@ -1,13 +1,16 @@
-package com.techcourse;
+package com.techcourse.servlet;
 
-import com.techcourse.controller.*;
+import com.techcourse.controller.LoginController;
+import com.techcourse.controller.LoginViewController;
+import com.techcourse.controller.LogoutController;
+import com.techcourse.controller.RegisterController;
+import com.techcourse.controller.RegisterViewController;
+import java.util.HashMap;
+import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class ManualHandlerMapping {
