+++ app/src/main/java/com/techcourse/servlet/UncheckedServletException.java
@@ -1,8 +1,8 @@
-package com.techcourse;
+package com.techcourse.servlet;
 
 public class UncheckedServletException extends RuntimeException {
 
-    public UncheckedServletException(Exception e) {
+    public UncheckedServletException(final Exception e) {
         super(e);
     }
 }
