+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler/HandlerKey.java
@@ -1,8 +1,7 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
-
-import web.org.springframework.web.bind.annotation.RequestMethod;
+package webmvc.org.springframework.web.servlet.mvc.handler;
 
 import java.util.Objects;
+import web.org.springframework.web.bind.annotation.RequestMethod;
 
 public class HandlerKey {
 
@@ -23,10 +22,14 @@ public class HandlerKey {
     }
 
     @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof HandlerKey)) return false;
-        HandlerKey that = (HandlerKey) o;
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof HandlerKey)) {
+            return false;
+        }
+        final HandlerKey that = (HandlerKey) o;
         return Objects.equals(url, that.url) && requestMethod == that.requestMethod;
     }
